feature: 1. fix some problem. 2. add front makefile logic.

+++ cmd/auth-server/types/types.go
@@ -228,7 +228,7 @@ type AttrValueResource struct {
 
 // ListInstanceResult list instance result.
 type ListInstanceResult struct {
-	Count   uint32             `json:"count"`
+	Count   uint64             `json:"count"`
 	Results []InstanceResource `json:"results"`
 }
